chore: rename `active filters` to `applied filters`
