+++ web/ee/components/issues/filters/applied-filters/issue-types.tsx
